Refactor EnvironmentValueMacro to match the rest of macro pattern
