+++ Sources/SwiftEnvironmentMacro/EnvironmentValueMacro.swift
@@ -20,66 +20,16 @@ public struct EnvironmentValueMacro: MemberMacro {
               extensionDeclaration.extendedType.as(IdentifierTypeSyntax.self)?.name.text == "EnvironmentValues" else {
             throw EnvironmentValueMacroError.attachedToInvalidType
         }
-        return extensionDeclaration.environmentDeclaration
+        let isPublic = extensionDeclaration.modifiers.contains { $0.trimmedDescription == "public" }
+        return extensionDeclaration.memberBlock.members
+            .staticVariables.toEnvironmentDeclarations(isPublic: isPublic)
             .map { "\(raw: $0)" }
-    }
-}
-
-struct EnvironmentDeclaration: CustomStringConvertible {
-    let isPublic: Bool
-    let baseName: String
-    let type: String
-    
-    var derivedName: String {
-        "\(baseName.firstCapitalized)SwiftEnvironmentKey"
-    }
-    var accessModifier: String {
-        isPublic ? "public " : ""
-    }
-    
-    var description: String {
-        """
-        struct \(derivedName): EnvironmentKey {
-            \(accessModifier)static let defaultValue: \(type) = EnvironmentValues.\(baseName)
-        }
-        
-        var \(baseName): \(type) {
-            get {
-                self[\(derivedName).self]
-            }
-            set {
-                self[\(derivedName).self] = newValue
-            }
-        }
-        """
     }
 }
 
 // MARK: Private extensions
 
-private extension ExtensionDeclSyntax {
-    var environmentDeclaration: [EnvironmentDeclaration] {
-        let isPublic = self.modifiers.contains { $0.trimmedDescription == "public" }
-        return memberBlock.members
-            .compactMap { $0.decl.as(VariableDeclSyntax.self) }
-            .filter { $0.isStatic }
-            .compactMap { variable in
-                guard let name = variable.name,
-                      let type = variable.typeAnnotation ?? variable.initializer else {
-                    return nil
-                }
-                return EnvironmentDeclaration(isPublic: isPublic, baseName: name, type: type)
-                
-            }
-    }
-}
-
 private extension VariableDeclSyntax {
-    var isStatic: Bool {
-        modifiers.contains {
-            $0.name.text == "static"
-        }
-    }
     
     var name: String? {
         bindings.first?.pattern
@@ -98,5 +48,17 @@ private extension VariableDeclSyntax {
             .as(FunctionCallExprSyntax.self)?
             .calledExpression
             .trimmedDescription
+    }
+}
+
+extension Sequence where Element == VariableDeclSyntax {
+    func toEnvironmentDeclarations(isPublic: Bool) -> [EnvironmentDeclaration] {
+        compactMap { variable in
+            guard let name = variable.name,
+                  let type = variable.typeAnnotation ?? variable.initializer else {
+                return nil
+            }
+            return EnvironmentDeclaration(isPublic: isPublic, baseName: name, type: type)
+        }
     }
 }
